// types/ProviderApplication.tsx
export interface ProviderApplicationData {
  fullName: string;
  cpf: string;
  birthDate: string;
  gender: string;
  whatsapp: string;
  phone: string;
  serviceName: string;
  hourlyRate: string;
  paymentMethods: string[];
  description: string;
  street: string;
  number: string;
  complement: string;
  neighborhood: string;
  city: string;
  state: string;
  zipCode: string;
  useCurrentLocation: boolean;
}